+++ src/backend_cairo/mod.rs
@@ -373,10 +373,10 @@ fn _calc_node_bbox(
         usvg::NodeKind::Path(ref path) => {
             Some(utils::path_bbox(&path.segments, path.stroke.as_ref(), &ts2))
         }
-        usvg::NodeKind::Text(_) => {
+        usvg::NodeKind::Text(ref text) => {
             let mut bbox = Rect::new_bbox();
 
-            text::draw_blocks(node, opt, cr, |block| {
+            text::draw_blocks(text, node, opt, cr, |block| {
                 cr.new_path();
 
                 let context = text::init_pango_context(opt, cr);